Create Subject folder Move Subject model and tests into its own folder. Update tests to mock the root store and client without throwing errors.
