+++ packages/lib-classifier/src/store/Subject/index.js
@@ -1,0 +1,1 @@
+export { default } from './Subject'
